requirements.txt
Django==3.1.3
djangorestframework==3.12.2
django-cors-headers==3.5.0
djongo==1.3.6
pymongo[tls,srv]==3.12.3
python-dotenv==1.0.0
django-environ==0.10.0